ikiAI_Agents/agents/chat_interface_agent.py
from flask import Flask, request
from config import Config
import chroma_db

app = Flask(__name__)
app.config.from_object(Config)

db = chroma_db.ChromaDB(app)

class ChatInterfaceAgent:

    def __init__(self, user_id):
        self.user_id = user_id

    def start_chat(self):
        # Implementation of chat initiation goes here
        pass

    def provide_task_guidance(self, message):
        # Implementation of task guidance goes here
        pass

    def store_chat_history(self, chat_data):
        # Store chat history in Chroma DB
        db.insert('chat_history', {'user_id': self.user_id, 'chat_data': chat_data})

@app.route('/start_chat', methods=['POST'])
def start_chat():
    data = request.get_json()
    agent = ChatInterfaceAgent(data['user_id'])
    agent.start_chat()
    return {'status': 'Chat started'}

@app.route('/provide_task_guidance', methods=['POST'])
def provide_task_guidance():
    data = request.get_json()
    agent = ChatInterfaceAgent(data['user_id'])
    agent.provide_task_guidance(data['message'])
    return {'status': 'Task guidance provided'}

@app.route('/store_chat_history', methods=['POST'])
def store_chat_history():
    data = request.get_json()
    agent = ChatInterfaceAgent(data['user_id'])
    agent.store_chat_history(data['chat_data'])
    return {'status': 'Chat history stored'}
